Updated map format + faster map compression + map updates
Firstly, I fixed some major bugs with the map format. Previously, the map would receive a u32, and then convert that to an f32. This had problems with precision, as well as JavaScript numbers being really annoying to work with, as they're all 64 bit floating points. I eventually fixed this by just converting the float into bytes, and then receiving that, leading to much better precision and everything being the exact size it's supposed to be. Doing this in custom.js was kind of a hack, however, as I had to convert the float into a Float32Array, then convert that into a Uint8Array. It works, though, so that's nice.

Secondly, I pushed a PR to the rust-brotli github repo requesting that they switch to packed_simd_2, since packed_simd doesn't build and isn't maintained. IN the meantime while I wait for them to merge it, I'm just using my PR in the mean time. I also heavily lowered the compression quality for debug builds, so that it will compress much quicker (though it's the highest it can be for release builds).

Finally, I made the maps look nicer (I think) by replacing the ugly leaf texture with an actual wall color.