Fixed health/death bug + opts
Firstly, I finally fixed both the health bug (health varying between net
clients) and the death bug (deaths not syncing between clients if the
damage message didn't arrive). Basically, instead of sending Damage
mwssages that have the possibility of failing, we just send the health
of the player and a DamageSource variable in the stat messages. The
DamageSource tells the last source of damage given to the player.
Basically, the player's health is just set to whatever the damage
message says it is, and then if the health is 0, the player dies and a
point is awarded to the player ID in the damage source.

Also did some helper_functions optimizations. I also tried to reduce
heap allocationa and perform general optimizations where O could.