(Mostly) Fixed #60, improved efficiency
So I worked for a long time on trying to fix #60. I at first thought it
might have been an issue with taking the average size of the player,
instead of just taking the size directly. To fix that, I made the x
collision search and y collision search seperate, then combined at the
end of the function, which actually helps to improve both the accuracy
of collision checks (since it takes the exact size of the object instead
of taking an average) and it alos helps with improving the efficiency,
since it means it does less iterations over all. Switching to an any()
instead of a map().reduce() also helped to reduce the number of overall
checks, which further helped with efficiency. Unfortuanetly, however,
this still wasn't enough to fix the issue (although it made it more
difficult to hit said bug).

Next up, I tried to figure out where exactly the bug was being hit.
After changing the any() to a find(), I found that the bug was only
triggered when a player's collision was at x == Some(1) and y ==
Some(1). I then tried to make it so that if a collision was triggered at
those coordinates (in the range iterator, the collision was nullified.
This had theo wonderful effect of making it so players would, instead of
being stuck, they'd just move through the wall.

Finally, I decided to even further improve the speed and accuracy of the
program by only doing iterative collisions when the distance was greater
than the size of the object, and that finally improved collision speeds
even more and fixed the bug (when players aren't moving very very fast)