Added save data, fixes #34
So I had to write a wrapper function in config.rs, and have multiple ways of saving methods.

For WASM, I needed to write a JS function to write and read to localStorage. For native, I used directories-next to find config directories for multiple OS's, and just write and read to the files (using the file name as a key).

So in order to make it easier to write and read any struct without implementing annoying string methods, I use the ron libary alongside serde and a generic function I wrote to automatically convert types that implement Serialize. For Deserialize (get_data), I got tired of dealing with lifetimes, so you just have to call from_str whenever you get the string.

I'm currently using the save data for keybindings, and for the player
customization. I haven't had the time to test it for native, but I can
confirm it works for WASM.