Genericized the get_data function
It was annoying, but I did eventually get it to work. Now you can just get and write any data structure, with a string as a key, effortlessly.

Also added to compile for the native cpu for basic-http-server and wasm-bindgen- cli, just for some quick addtions to how fast the program runs.