Fixed players shoot many bullets bug
Added an if statement to not broadcast net shot messages that aren't
sent by the current player. (If we receive a shoot event that's from
another player, we shouldn't send it again).

Ladies and gentleman, I'm the greatest programmer of all time. This
incredibly massive change took 99% of my brain power, but I did it.