Fixed WASM screen scaling + config is own crates
Firstly, I made it so that the screen will scale depending on the
resolution of the user's screen. This will slightly stretch images, but
in exchange, the player's viewport isn't smaller or bigger depending on
the size of the player's screen. Will probably change this later to be
better.

Secondly,I moved the config.rs file into its own seperate crate, for
faster parallel compiles ofc.