Fixed cloak+loading, inferno buff, modified map
Firstly, I fixed the cloak ability by having players send their color
alpha values with each stat packe they send, since before, their alpha
value was constantly being reset.

Secondly, while playing with Inferno against @Susorodni, I noticed that
it was relatively difficult to both hit them with the molotov liquid,
and to light them on fire long enough to deal any useful amount of
damage. To compensate for this, I made it so that when hit by the
molotov liquid, the player's running speed was slowed to around 65% of
the usual speed for 2 seconds. This should hopefully make sucessfully
lighting a player on fire much easier.

I also fixed a bug where players' abilities wouldn't load properly at
the start, since while the ability resource was being set, the actual
player's character was not. This applies to perk and gun models as well.

Finally, I modified the map a bit to make it a little more fun to play
on.