Better resolution scaling
So now, isntead of stretching the screen in some weird way, the game
will always run at  1366x768 scale.The game, on startup, first
calculates the largest way that scale could be achieved, then, it
changes the actual camera scale so that no matter how big or small the
screen is, every player will have the exact same view of the map. Of
course, for very small screens, the game renders text and buttons at
too large a font, but that's an issue for another day. (Only tested on
WASM so far).