Added player timeouts, fixes #54
So basically, I just changed from a BTreeSet to a (single byte), which
not only stores the player's ID, but also the player's network handle
if it exists (with an obvious exception for the player hosting the
game). It also has a timer that shows the amount of time since the last
packet. Then, if a player doesn't send any packets for 15 seconds,
they're completely removed from the game. Unfortuanetly, the
disconnected player won't know they're disconnected, so we should
probably have a DisconnectedPlayers hashmap in order to send them a
command saying they've timed out.