Bullet collision, damage, speedball, faster cargo-make
So this commit has taken a little longer than I expected, mostly because I spent a long time trying to understand Rapier better. So bullets can now collide with walls and players, and when they do, they'll destroy themselves.

Players can now also take damage. I thought about how to architect this for a little while, since I don't want to repeat the mistakes made in the previous logic.rs function where it loops multiple times. I eventually just decided to abuse the Option enum a lot, and just use said optionals to decide whether it's a bullet or a player, or something entirely different. The damage a bullet can do is stored as user_data in the rigid body, using a fancy function I wrote to store an f32 as a u128, and reverse said transition.

Speedball's also work, in the sense that they speed up and get bigger over time. They unfortunately don't currently increase how much damage they do over time, but it won't be difficult to make it so that they will. I specify this as one of the changes made to show that the physics engine is pretty powerful, and that by just making the linear damping negative, and adding a cap to how fast the linear velocity can be, we can recreate the speedball in very few lines of code.

I quickly made it so that the game won't crash if you don't have an internet connection on start up, but added a warning that it's basically a local game that only works on Chromium based browsers.

Finally, I added some config options to cargo-make to make it run faster, though you need to update to the absolute latest version.