Added Damage + Death + Multip. Movem.
Pretty big step in getting the full game working is to add player damage and deaths. Now, an entire game can be played.

The reason this commit took so long to make was trying to figure out how to send the shot_from aspect of bullets in rigid bodies. This wouldn't be an issue in the old game function, but in this one, I refuse to 1. loop twice through the query for any reason whatsoever, and 2. have multiple different queries that do two very similar things. An issue that I had with the old move_objects fn was that it was very slow and inefficient due to those two things. I finally figured out how to do this by just including it in the user_data, which is starting to get dangerously full (just 4 bytes left). This is a workable solution for now, and as long as I don't need to send more data in weird ways, it should work.

Also fixed multiplayer movement, since I was writing to the sprite translation instead of the rigid body translation, so it would just get overwritten constantly. Maybe I should turn enemy players into RigidBody::Static for efficiency?