Pulsewaves
So once again, I was thinking about how to efficiently (no double looping) apply the slow down effect to players, and of course, packing it into the u128 was the answer. I just pack an extra byte in saying the type of projectile. I also wanted to make sure that players had a user_data (so that pulse waves only destroyed themselves if they hit a player), so the player user_data correspond to u128::MAX, since the chances of that naturally occurring are 0, and will only become possible if every part of the u128 is full, and at max value, which is very unlikely to happen.

Also fixed a minor bug with spawning projectiles, where sometimes they would spawn inside of the player and immediately despawn. To counteract this, projectiles move 1 frame at the beginning when they're spawned.