Fixed Widowmaker
So in order to be able to heal across frames, without looping over the query twice, I crated a WidowMakerHeals struct that contains all the players that need to be healed since their widowmaker bullets hit someone. To do this, I also created the WidowMaker projectile that has some special attributes to the game.

I also made it so that every time the WidowMaker is fired, the player shooting it loses health equivalent to how much damage the gun does.