Bot Trait + AggroBot
The Bot Trait allows any struct with any sort of internal structure to become a bot, as long as it follows some basic structures. Firstly, upon being initialized, every bot is given the location of all bots, how the map looks, and said bot's player id. The player id is only ever sent once, so it's recommended to store it inside the struct. The rest of the required functions are for updating the bots' internals due to the players or map, miscellaneous changes the bot wants to make to itself. Finally, some data is received from the bot on where it wants to move, where it wants to face, whether it wants to use it's ability or not, and whether or not it should attempt to shoot.

Instead of being fed the player struct itself, which contains a lot of information that bots shouldn't know, the bot is given a TruncatedPlayer, which contains minimal information about the player. Currently, it only contains the player's position and ID, but later it will include stuff like it's rotation.

Tehe AggroBot is a relatively simple example of a bot that tries to move
towards players if they get close-ish, and get within shooting range.
When I finish up the shooting part of bots, it will fire at them.