General refactoring
I removed (almost) all of the warnings in all of the subcrates and game library and binary, as well as fixing all of the clippy complaints as well. This should also help to improve performance since we're not getting unnecessary data in stuff like queries, and there was evena n unnecessary collision check in the Ghost ability.