Updated to latest Bevy git commit
So normally, I just run `cargo update` every couple commits and that's
it, but however, this time it's different. Bevy made it so that it
doesn't automatically implement the Component trait, so it has to be
manually implemented for every type. The best part is that it doesn't
do it for external types either, so we need to make a ton of wrapper
types for literally every single external library type we use.