Fixed players moving through walls, better warp, grenade nerf, chat fix
Firstly, I fixed players moving through walls and improved the warp ability at the same time. Essentially, I enabled CCD permanently on players (for the walls stuff) and made it to where the warp physically moves the player instead of improving the velocity by a ton.

I nerfed the grenade as well, since i found it to be extremely oppressive even with the 3 second delay for explosions.

Finally, I made a quick fix to text chat to make it to where you can only send 1 character in chat. This should fix most of the weird special character stuff you could send in chat. I also added a small if statement so that players could send numbers in chat as well.