Fixed and buffed Brute
Closes #68

So the Brute now uses forces, making it so that it actually works properly with the physics engine. These forces are only applied if the player is local as well, making it so that weird twitching that would happen if a brute hit you with their ability doesn't happen anymore.

The brute's recharge rate is too fast rn imo, but I want to see some gameplay before applying any nerfs needed.