Dead players collision fix + more spawn points
So as an optimization / out of laziness, the game creates players before they join in, and then just makes the players visible when a player joins. However, an issue with this is that it was possible to interact with these players (shoot the, move them, etc). Players who died also had this issue, as they'd move and could get hit by bullets. To counteract this, I made it so that players would be a part of no InteractionGroups when they died, and then got set back to normal when they respawned.