Added lighting shaders to the game
That's right ladies and gentlemen, it's finally here.

First off, I could rant for so long about how much I absolutely despise GLSL, and how frustrating it was to get it to work (and as you'll read later, it still isn't exactly being an angel).

Now, what did I actually do? All lights are stored in an ArrayVec called LightsRes. Whenever any system wants to add or remove a light from the game, they should push to this ArrayVec and store the index at which their light is being held, then remove said index when they're finished? So to actually pass around all the lights to all of the different entities that are affected by lights (almost every entity with a sprite), we need to copy lights_res into each individual components Lights struct (which is the struct that is actually used in GLSL). This is the reason I use ArrayVec. If this were to be stored in a regular Vec, we could save on quite a bit of storage space, but in return, we would spend a lot of time copying from the heap to the stack. After some deliberation, I thought it would be faster to just store all of the data in the stack, and just do small copies.

Finally, we get to the good stuff. GLSL is a demonic language that I unfortunately have to work with. I don't want to go into detail about the add_lighting function, but it basically just makes objects darker depending on their distance from each light source. The one in the player.frag is a buggy one I kept bc it makes players look shiny, and that's cool.