Performance improvements
I've been trying for the past 3 days to get lights to appear when you shoot, and it isn't working. I've gotten many performance improvements within that span of time though, so I'll just include those. It's just a otn of refactoring. It definetly isn't an issue with the LightsResource structure, or the lights.value structure, since those work perfectly after VERY extensive testing.