Full TCP Client and Server
That's right, the actual networking code and bincode code is finally complete! The best part is that bincode is actually used internally within turbulence (part of bevy_networking_turbulence) so all of our types are already serializable by default! THe only thing left is to implement it within the game itself (hopefully).

I also split the crate into 3 for quicker compile times and to reduce code duplication.