Mold linker for debug builds + faster compression
I've been looking at using the mold linker for debug builds in the
project for a while, but it was always just a little bit too buggy to
build consistently. However, I've been able to successfully begin
building with the linker for a while now, so I'm switching it to be the
default debug linker. In release builds, LLD will continue to be used,
since mold doesn't currently support LTO, thin or fat.

I also made it so that instead of using rust-script, the Makefile just
uses an equivalent tool I cooked up in tools/compress_wasm. It should
actually make compressing wasm faster since, firstly, you don't need to
recompile the script every single time you wanna use it (it's just a one
time and never again tool). It also runs with a lot more optimization
and with march=native since it will really only run locally.