Fixed native typing bug
When typing out the IP address on native builds when joining a game, I guess for some reason the backspace character would be added while typing or something? I don't know, maybe it was something else, but either way I fixed it by adding a check to make sure all characters typed in are numbers or the "." character.