Fixed player joining multiple times bug
So before, if a player sent multiple ability requests before the server
could actually send back the ability, the game would spawn a player for
each request, which would usually lead to a fun crash or weird gameplay.
It was also possible for players to send multiple ID requests before
receiving a response, which would lead to no more ID's being available,
and players being in a weird limbo state where they're both joined and
not joined. That should all be fixed now with some actual checks before
letting a player join.