Name visibility bug fixes + set ability crash fix + new player crash fix
Firstly, I would've commited this to main then rebased normally, but one
of the bug fixes is dependent on changes made in the TCP branch so it
makes more sense to commit it here.

Anyway, so firstly, I fixed a bug where when a player died, their name
would stay visible, as well as fixing a bug where the player's name
would stay visible while the player was cloaking.

Secondly, I fixed a crash where the game would crash if the player tried
to change their ability in the middle of a game (I was sending to the
wrong message channel).

Finally, I fixed a crash that would occur when a new player would join
the game after a previous one disconnected. This was because I was
completely despawning the player, as opposed to just turning them
invisible, which would mean that when a player tried to control the new
entity, it didn't exist.

Fixes multi player joining crash issue on #69